Enable linting of CMake files using pre-commit (#9484)
This PR resolves #9396, making it easy to run `cmake-format` and `cmake-lint` on CMake files in cudf using pre-commit. The one additional complexity associated with these hooks compared to our other pre-commit hooks is the need for the `cmake-format-rapids-cmake.json` configuration file that is maintained in [rapids-cmake](https://github.com/rapidsai/rapids-cmake) and is only added during the build process. We don't want pre-commit to fail outright when the file isn't present, so this PR introduces a custom script wrapping the `cmake-format` and `cmake-lint` hooks in a check for the presence of the file. The script respects an associated environment variable that advanced users can exploit if they store this file in a nonstandard location.

Once this PR is finalized I'll make a follow-up PR that actually applies the formatting to our existing CMake.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/cudf/pull/9484