Adds cudaProfilerStart/cudaProfilerStop in JNI api (#9543)
Closes #9399 

This PR adds support for `cudaProfilerStart` and `cudaProfilerStop`. I did not add a test for this case as I don't think it makes sense, given that a user could be legitimately capturing a profile of the tests (so a start/stop could change things).

The way one can use this with `nsys` is:

```
nsys profile -c cudaProfilerApi --capture-range-end=repeat
```

Where there are other options for what to do on capture end according to [the docs](https://docs.nvidia.com/nsight-systems/UserGuide/index.html). The command above stops collecting, and restarts again when `cudaProfilerStart` is invoked. The nsys profiled process doesn't need to be restarted, which is useful if one is trying to capture several experiments back-to-back.

There is a way to do this with nvtx ranges triggering the capture, but I could not get it to work with named ranges. I was told to try registering the name of the nvtx range, and we can still try this, but I felt the `cudaProfiler*` api was useful on its own (and for our purposes).

Authors:
  - Alessandro Bellina (https://github.com/abellina)

Approvers:
  - Jim Brennan (https://github.com/jbrennan333)
  - Robert (Bobby) Evans (https://github.com/revans2)
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/9543