Remove sizeof and standardize on memory_usage (#9544)
This PR removes implementations of `__sizeof__` from cudf classes. Previously, `__sizeof__` was overridden to return the total GPU memory usage, but this is inconsistent with the standard Python semantics of this function and should be removed. The appropriate way to query for total GPU memory usage is via the `memory_usage` function, which is now standardized across various objects. The sizeof dispatch for dask is set to use `memory_usage` as well to avoid any breakage here.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)
  - Charles Blackmon-Luca (https://github.com/charlesbluca)
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/cudf/pull/9544