Correct `to_cupy`/`to_numpy` order (#9688)
If I'm not mistaken, before this patch, the new functions `to_cupy`/`to_numpy`  are creating arrays with the wrong/unexpected order  'C' (given that is not configurable) as opposed to 'F' major corresponding to cuDF's column ordering. This is very important for cuML interoperability since it makes cuML do conversions unnecessarily.