Fix an out-of-bounds read in validity copying in contiguous_split. (#9842)
Fixes https://github.com/rapidsai/cudf/issues/9504 

The bug was pretty straightforward:  when copying validity bits, we (potentially) perform a bit shift on the data so that we can read aligned 4 bytes at a time.  Under certain circumstances, we were reading 1 word past the end of the input incorrectly.

Adding a do not merge tag - waiting to get a full run of TPC-DS with this as an extra safety check.

Authors:
  - https://github.com/nvdbaranec

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - David Wendt (https://github.com/davidwendt)

URL: https://github.com/rapidsai/cudf/pull/9842