Fix JNI leak on copy to device (#10229)
We were not closing intermediate column view handles when copying nested data from the device to the host. I think we could completely replace `NestedColumnVector` with `ColumnView` in the future, but it would require some optimizations in `ColumnView` to make it happen cleanly, so I didn't do any of that here.

I also renamed one of the methods that had a leak so it would be more clear what it was doing, and I marked it as private so it would be less likely to leak out and be used by other code.

Authors:
  - Robert (Bobby) Evans (https://github.com/revans2)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/10229