Limiting async allocator using alignment of 512 (#10395)
We use the `limiting_resource_adaptor` in front of the async allocator to track the total size of allocations and trigger spills. By default the adaptor aligns to 256, but the async allocator internally aligns to 512. This causes the limiting adaptor to under count allocation sizes, leading to OOMs in the async allocator. This PR changes the limiting adaptor to also use 512. This seems to help with a customer query that we were seeing OOMs before.

Note that CUDA is planning to switch back to 256 in the future so this may need to be revisited.

Fixes #10384

Authors:
  - Rong Ou (https://github.com/rongou)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/10395