Fix lifespan of the temporary directory that holds cuFile configuration file (#10403)
The temporary directory is created when we modify the cuFile JSON file to control the compat mode. This directory is a local variable in the `modify_cufile_json` function, so the modified JSON file is deleted as soon as the function returns.
Because of this, cuFile is not able to read the modified JSON config and falls back to the default config file.

This PR makes the temporary directory a static variable so that the cuFile config file is avaliable until the end of program. 
This fixes the compat mode override that we do via `LIBCUDF_CUFILE_POLICY` environment variable.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/cudf/pull/10403