Fix for integer overflow in contiguous-split (#10437)
This is a fix for an integer overflow in `contiguous_split.cu::copy_data`. 

We encountered this problem when running a large query with the spark-rapids plug-in.  The plug-in OOM callback was reporting negative numbers for some failed allocations.  It treats the allocation size as a `long`.   The allocations were coming from `contiguous_split` with a large number of rows (399,000,000) of type `UINT64`.

With help from @nvdbaranec, I was able to isolate the problem to this line:
```
size_t const bytes = buf.num_elements * buf.element_size;
```

The two operands are both `INT32`.  So the result of the multiply is being sign-extended when assigning to `bytes`.

The fix is to cast these to `size_t` before the multiply.

Authors:
  - Jim Brennan (https://github.com/jbrennan333)

Approvers:
  - https://github.com/nvdbaranec
  - Yunsong Wang (https://github.com/PointKernel)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/10437