Allow building arrow with parquet and not python (#10702)
Not 100% sure if this would classify as a bug or an improvement. This updates the cmake so I can build with parquet support in arrow, but not build for python.  This just moves some configs down from the python support to parquet support. From what I have seen the python code builds with parquet enabled, and we want to start to experiment in Spark with using the native parquet footer parsing to improve performance on files with lots of columns.

I am happy to duplicate the configs if there is the option to build python support without parquet and it shows problem.

Authors:
  - Robert (Bobby) Evans (https://github.com/revans2)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/cudf/pull/10702