Use canonical "magic methods" (replace `x.__repr__()` with `repr(x)`). (#10735)
This PR uses the canonical builtin function `repr(x)` to compute representations instead of the double-underscored form `x.__repr__()`. The magic method `def __repr__(self): ...` is how a custom representation is defined for a class, but it is not the typical way to call that function. Same for `x.__iter__()` :arrow_right: `iter(x)` and similarly for `__hash__`, `__str__`, etc. This PR fixes all of the instances of that pattern that I could find, mostly affecting `repr(x)`.

Note that non-standard magic methods like `__dlpack__` or `__dataframe__` that are not defined in the Python object model do not necessarily follow this convention where a free function of the same name is used to call those methods.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/10735