Implement DataFrame.eval using libcudf ASTs (#8022)
This PR exposes `libcudf`'s expression parsing functionality in `cudf` and uses it to implement `DataFrame.eval`. The implementation is mostly feature-complete, but there are a few limitations relative to the `pandas` API and a couple of gotchas around type casting. The implementation is reasonably performant, improving upon an equivalent `df.apply` even accounting for JIT-compilation overhead. This implementation provides a stepping stone to leveraging `libcudf`'s AST implementation for more complex tasks in `cudf` such as conditional joins.

The most significant issue with the current implementation is the lack of casting between integral types, meaning that operations can only be performed between columns of the _exact_ same dtype. For example, operations between int8 and int16 would fail. This becomes particularly problematic for constants e.g. `df.eval('x+1')`. The best paths to improve this are at the C++ level of the expression evaluation, so I think we'll have to live with this limitation for now if we want to move forward.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Ashwin Srinath (https://github.com/shwina)
  - Bradley Dice (https://github.com/bdice)
  - https://github.com/brandon-b-miller

URL: https://github.com/rapidsai/cudf/pull/8022